moving class to outside derivative package
